requirements.txt
bs4
python-dotenv
langchain
langchain_community
requests
streamlit
